Correct installation location and improve code
Installing something like this into `/bin/` would be _very_ bad
practice. A more appropriate place would `/usr/bin/`, although another
common one is, I think, `/opt/`. I've addressed that in this commit.

I've added some color. Just makes it look nicer, but of course not
everyone will or can see it or see it properly. Feel free to get rid of
the color escape sequences. Using `tput` to set the colors is better
practice (more portable), but I have issues with that personally, so I
wasn't sure if other people would too.

Used `Exec()` to just avoid code repetition, and to make the whole
process cleaner and easier with which to work.

Errors are now a thing, and it's now much more robust.

The ownership and group are now appropriately set, given its target.

Lastly, ensured correct mode settings, because it's very important this
is set correctly when installing to a PATH location. 755 is the norm.